requirements.txt
pytest
selene>=2.0.0b17
appium-python-client
allure-pytest
python-dotenv